Cleaned up building code. ready for merge
