+++ server/src/router/buildings.ts
@@ -1,6 +1,6 @@
 import express from 'express';
 import { db } from '../firebase';
-import { BuildingType, RoomIDType, RoomType } from '../types';
+import { BuildingType } from '../types';
 
 const router = express.Router();
 
@@ -34,15 +34,6 @@ router.get("/:b_id", async (req, res) => {
 	res.send(data);
 });
 
-router.put("/:b_id/buildings/1", (req, res) => {
-	res.json({ message: "Added and Updated Building" })
-});
-
-router.put("/updateBuilding/buildings/:b_id/room/:r_id", (req, res) => {
-	res.json({ message: "Created New Rooms!" })
-});
-
-
 /***
  *  Removes a building by its ID
  */
@@ -51,6 +42,7 @@ router.delete("/:b_id", async (req, res) => {
 	const buildingsCollection = db.collection('buildings');
 	const ref = buildingsCollection.doc(buildingID);
 	ref.delete();
+	res.send(`Deleted building ${buildingID}`);
 });
 
 
@@ -59,7 +51,7 @@ router.delete("/:b_id", async (req, res) => {
  */
 router.post("/", async (req, res) => {
 	const buildingsCollection = await db.collection('buildings');
-	const ref = buildingsCollection.doc(req.body.b_id);
+	const ref = buildingsCollection.doc(req.body.b_id.toString());
 	const building: BuildingType = {
 		b_id: req.body.b_id,
 		description: req.body.description,
@@ -71,48 +63,5 @@ router.post("/", async (req, res) => {
 	await ref.set(building);
 	res.send(building);
 });
-
-
-/***
- * Adds a new room with the next available ID (i hope it does LOL)
- */
-router.post("/:b_id/rooms/", async (req, res) => {
-	const buildingID = req.params.b_id;
-	const buildingsCollection = db.collection('buildings');
-	const ref = buildingsCollection.doc(buildingID);
-
-	const doc = await ref.get();
-	const rooms: RoomIDType[] = await doc.get('rooms');
-	const room: RoomType = req.body;
-
-	const updated_rooms = [...rooms, room];
-	res.send(updated_rooms);
-	await ref.update({ rooms: updated_rooms });
-	res.send(ref);
-});
-
-/***
- * Returns information about a room (i hope it does too LOL)
- */
-router.get("/:b_id/rooms/:r_id", async (req, res) => {
-	const roomID = req.params.r_id;
-	const roomCollection = db.collection('rooms');
-	const ref = roomCollection.doc(roomID);
-	const doc = await ref.get();
-	const roomsInfo: RoomType = await doc.get('rooms');
-	res.send(roomsInfo);
-});
-
-
-/***
- * Removes a room by its ID (also really hoping this works)
- */
-router.delete("/:b_id/rooms/:r_id", async (req, res) => {
-	const roomID = req.params.r_id;
-	const roomsCollection = db.collection('rooms');
-	const ref = roomsCollection.doc(roomID);
-	ref.delete();
-});
-
 
 export default router;
